Finding determinant was made.

+++ lab_08_16/inc/alloc_free_matrix.h
@@ -5,6 +5,6 @@
 #include <stdio.h>
 #include "errors.h"
 
-int **alloc_matrix(size_t rows, size_t cols);
+double **alloc_matrix(size_t rows, size_t cols);
 
 #endif // __ALLOC_FREE_MATRIX_H__
