+++ lab_08_16/src/alloc_free_matrix.c
@@ -1,14 +1,14 @@
 #include "../inc/alloc_free_matrix.h"
 
-int **alloc_matrix(size_t rows, size_t cols)
+double **alloc_matrix(size_t rows, size_t cols)
 {
     // int **data = malloc(rows * sizeof(int *) + rows * cols * sizeof(int));
-    int **data = calloc(rows * sizeof(int *) + rows * cols, sizeof(int));
+    double **data = calloc(rows * sizeof(double *) + rows * cols, sizeof(double));
     if (!data)
         return NULL;
     
     for (size_t i = 0; i < rows; i++)
-        data[i] = (int *)((char *)data + rows * sizeof(int *) + i * cols * sizeof(int));
+        data[i] = (double *)((char *)data + rows * sizeof(double *) + i * cols * sizeof(double));
     
     return data;
 }
