+++ lab_08_16/src/file_handler.c
@@ -31,13 +31,13 @@ size_t file_first_line(FILE *f, size_t *rows, size_t *cols, size_t *non_zero, si
     return EXIT_SUCCESS;
 }
 
-size_t fill_matr_from_file(FILE *f, int **matrix, size_t rows, size_t cols, size_t non_zero, size_t input_type)
+size_t fill_matr_from_file(FILE *f, double **matrix, size_t rows, size_t cols, size_t non_zero, size_t input_type)
 {
     if (input_type == 1)
     {
         for (size_t i = 0; i < rows; i++)
             for (size_t j = 0; j < cols; j++)
-                if (fscanf(f, "%d", &matrix[i][j]) != 1)
+                if (fscanf(f, "%lf", &matrix[i][j]) != 1)
                     return INCORRECT_INPUT;
     }
     else if (input_type == 2)
@@ -54,12 +54,12 @@ size_t fill_matr_from_file(FILE *f, int **matrix, size_t rows, size_t cols, size
     return EXIT_SUCCESS;
 }
 
-size_t file_mtr_write(FILE *f, int **mtr, size_t rows, size_t cols)
+size_t file_mtr_write(FILE *f, double **mtr, size_t rows, size_t cols)
 {
     for (size_t i = 0; i < rows; i++)
     {
         for (size_t j = 0; j < cols; j++)
-            if (fprintf(f, "%d ", mtr[i][j]) < 0)
+            if (fprintf(f, "%d ", (int)mtr[i][j]) < 0)
                 return WRITE_ERROR;
         fprintf(f, "\n");
     }
